@@ -4,7 +4,7 @@
 
     <groupId>de.quinscape</groupId>
     <artifactId>spring-jsview</artifactId>
-    <version>0.12-SNAPSHOT</version>
+    <version>1.0-SNAPSHOT</version>
 
     <name>Spring WebMVC JavaScript Views</name>
     <description>
@@ -34,11 +34,11 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <java.version>17</java.version>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
 
-        <spring.core.version>5.2.4.RELEASE</spring.core.version>
+        <spring.core.version>6.0.5</spring.core.version>
 
     </properties>
 
@@ -52,13 +52,13 @@
         <dependency>
             <groupId>com.google.code.svenson</groupId>
             <artifactId>svenson</artifactId>
-            <version>1.5.8</version>
+            <version>1.6.1</version>
         </dependency>
 
         <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
-            <version>2.4</version>
+            <version>2.7</version>
         </dependency>
 
         <dependency>
@@ -68,9 +68,9 @@
         </dependency>
 
         <dependency>
-            <groupId>javax.servlet</groupId>
-            <artifactId>javax.servlet-api</artifactId>
-            <version>3.1.0</version>
+            <groupId>jakarta.servlet</groupId>
+            <artifactId>jakarta.servlet-api</artifactId>
+            <version>6.0.0</version>
             <scope>provided</scope>
         </dependency>
 
