@@ -4,7 +4,7 @@ import de.quinscape.spring.jsview.util.Util;
 import org.svenson.JSONParameters;
 import org.svenson.JSONTypeHint;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.Map;
 
