@@ -3,7 +3,7 @@ package de.quinscape.spring.jsview.loader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.ServletContext;
+import jakarta.servlet.ServletContext;
 import java.io.InputStream;
 
 /**
